requirements.txt
pillow==12.0.0
ultralytics==8.3.228
